requirements.txt
google-generativeai>=0.3.0
minio>=7.2.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
aiofiles>=23.2.1
python-multipart>=0.0.6
